Fixed a compilation problem that occurred when compiling tests with Ant
git-svn-id: http://agilefant.googlecode.com/svn/branches/Monza_restructure@2441 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ test/fi/hut/soberit/agilefant/business/IterationBusinessTest.java
@@ -204,8 +204,8 @@ public class IterationBusinessTest {
         
         IterationMetrics actualMetrics = iterationBusiness.getIterationMetrics(iteration); 
         
-        assertEquals(0, actualMetrics.getPercentDoneStories());
-        assertEquals(0, actualMetrics.getPercentDoneStories());
+        assertEquals(0, actualMetrics.getPercentDoneStories().intValue());
+        assertEquals(0, actualMetrics.getPercentDoneStories().intValue());
         
         verify(iterationHistoryEntryBusiness, iterationDAO);
     }
